moved main and util to src directory
